+++ CHANGELOG.md
@@ -5,9 +5,9 @@ All notable changes for each version of this project will be documented in this 
 ## 10.0.0
 
 ### General
-- `IgxTransaction` - The `onStateUpdate` now emits with information of its origin. The emitted value is of type `IStateUpdateEvent`, which has two properies:
-    - `origin` - a mandatory property. It can vary within the values of the `TransactionEventOrigin` interface; 
-    - `actions` - an optional property that contains information about the transactions, that caused the emission of the event. It is used by `undo` and `redo`.
+- `IgxTransaction` - The `onStateUpdate` now emits with information of its origin. The emitted value is of type `StateUpdateEvent`, which has two properies:
+    - `origin` - it can vary within the values of the `TransactionEventOrigin` interface; 
+    - `actions` - contains information about the transactions, that caused the emission of the event.
 
 
 ## 9.1.1